# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SentimentLabel =
  (typeof SentimentLabel)[keyof typeof SentimentLabel];

export const SentimentLabel = {
  positive: "positive",
  negative: "negative",
  neutral: "neutral",
} as const;

export type ModelName = (typeof ModelName)[keyof typeof ModelName];

export const ModelName = {
  naive_bayes: "naive_bayes",
  logistic_regression: "logistic_regression",
  svm: "svm",
} as const;

export type SentimentResultScores = {
  positive: number;
  negative: number;
  neutral: number;
};

export interface SentimentResult {
  label: SentimentLabel;
  confidence: number;
  scores: SentimentResultScores;
  model: ModelName;
  processedText: string;
}

export interface AnalyzeTextRequest {
  text: string;
  model?: ModelName;
}

export interface AnalyzeTextResponse {
  originalText: string;
  result: SentimentResult;
  allModels: SentimentResult[];
  keywords: string[];
  processingTimeMs: number;
}

export interface AnalyzeBatchRequest {
  texts: string[];
  model?: ModelName;
  textColumn?: string;
}

export interface BatchTextResult {
  index: number;
  text: string;
  label: SentimentLabel;
  confidence: number;
  model: ModelName;
}

export type AnalyzeBatchResponseSummary = {
  total: number;
  positive: number;
  negative: number;
  neutral: number;
  positivePercent: number;
  negativePercent: number;
  neutralPercent: number;
};

export interface AnalyzeBatchResponse {
  results: BatchTextResult[];
  summary: AnalyzeBatchResponseSummary;
  processingTimeMs: number;
}

export type SampleDatasetResponseSamplesItem = {
  id: number;
  text: string;
  label: SentimentLabel;
  source: string;
};

export interface SampleDatasetResponse {
  samples: SampleDatasetResponseSamplesItem[];
  total: number;
  datasetName: string;
}

export interface TrainRequest {
  texts?: string[];
  labels?: SentimentLabel[];
  useDefaultDataset?: boolean;
}

export interface ConfusionMatrix {
  matrix: number[][];
  labels: string[];
}

export type ModelMetricsClassReport = {
  [key: string]: {
    precision: number;
    recall: number;
    f1Score: number;
    support: number;
  };
};

export interface ModelMetrics {
  model: ModelName;
  accuracy: number;
  precision: number;
  recall: number;
  f1Score: number;
  confusionMatrix: ConfusionMatrix;
  classReport: ModelMetricsClassReport;
  trainingSamples: number;
  testSamples: number;
  trainingTimeMs: number;
}

export interface TrainResponse {
  success: boolean;
  metrics: ModelMetrics[];
  bestModel: ModelName;
  totalTimeMs: number;
}

export interface ModelMetricsResponse {
  trained: boolean;
  metrics: ModelMetrics[];
  bestModel?: ModelName;
  lastTrainedAt?: string | null;
}

export interface CompareRequest {
  text: string;
}

export interface CompareResponse {
  text: string;
  comparisons: SentimentResult[];
  consensus: SentimentLabel;
  agreement: number;
}

export type GenerateReportRequestFormat =
  (typeof GenerateReportRequestFormat)[keyof typeof GenerateReportRequestFormat];

export const GenerateReportRequestFormat = {
  json: "json",
  csv: "csv",
} as const;

export interface GenerateReportRequest {
  results: BatchTextResult[];
  format: GenerateReportRequestFormat;
  includeMetrics?: boolean;
}

export interface ReportResponse {
  data: string;
  filename: string;
  format: string;
}

export interface WordCloudRequest {
  texts: string[];
  sentimentFilter?: SentimentLabel;
}

export type WordCloudResponseWordsItem = {
  text: string;
  value: number;
};

export interface WordCloudResponse {
  words: WordCloudResponseWordsItem[];
}

export type GetSampleDatasetParams = {
  limit?: number;
};