# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Analyze sentiment of custom text input using selected or all ML models
 * @summary Analyze sentiment of text
 */
export const AnalyzeSentimentBody = zod.object({
  text: zod.string(),
  model: zod.enum(["naive_bayes", "logistic_regression", "svm"]).optional(),
});

export const AnalyzeSentimentResponse = zod.object({
  originalText: zod.string(),
  result: zod.object({
    label: zod.enum(["positive", "negative", "neutral"]),
    confidence: zod.number(),
    scores: zod.object({
      positive: zod.number(),
      negative: zod.number(),
      neutral: zod.number(),
    }),
    model: zod.enum(["naive_bayes", "logistic_regression", "svm"]),
    processedText: zod.string(),
  }),
  allModels: zod.array(
    zod.object({
      label: zod.enum(["positive", "negative", "neutral"]),
      confidence: zod.number(),
      scores: zod.object({
        positive: zod.number(),
        negative: zod.number(),
        neutral: zod.number(),
      }),
      model: zod.enum(["naive_bayes", "logistic_regression", "svm"]),
      processedText: zod.string(),
    }),
  ),
  keywords: zod.array(zod.string()),
  processingTimeMs: zod.number(),
});

/**
 * Analyze sentiment of a batch of texts (from uploaded CSV)
 * @summary Analyze sentiment of multiple texts
 */
export const AnalyzeBatchBody = zod.object({
  texts: zod.array(zod.string()),
  model: zod.enum(["naive_bayes", "logistic_regression", "svm"]).optional(),
  textColumn: zod.string().optional(),
});

export const AnalyzeBatchResponse = zod.object({
  results: zod.array(
    zod.object({
      index: zod.number(),
      text: zod.string(),
      label: zod.enum(["positive", "negative", "neutral"]),
      confidence: zod.number(),
      model: zod.enum(["naive_bayes", "logistic_regression", "svm"]),
    }),
  ),
  summary: zod.object({
    total: zod.number(),
    positive: zod.number(),
    negative: zod.number(),
    neutral: zod.number(),
    positivePercent: zod.number(),
    negativePercent: zod.number(),
    neutralPercent: zod.number(),
  }),
  processingTimeMs: zod.number(),
});

/**
 * Returns pre-loaded sample tweets with their sentiment labels
 * @summary Get sample dataset
 */
export const getSampleDatasetQueryLimitDefault = 50;

export const GetSampleDatasetQueryParams = zod.object({
  limit: zod.coerce.number().default(getSampleDatasetQueryLimitDefault),
});

export const GetSampleDatasetResponse = zod.object({
  samples: zod.array(
    zod.object({
      id: zod.number(),
      text: zod.string(),
      label: zod.enum(["positive", "negative", "neutral"]),
      source: zod.string(),
    }),
  ),
  total: zod.number(),
  datasetName: zod.string(),
});

/**
 * Train all ML models on the provided dataset
 * @summary Train ML models
 */
export const TrainModelsBody = zod.object({
  texts: zod.array(zod.string()).optional(),
  labels: zod.array(zod.enum(["positive", "negative", "neutral"])).optional(),
  useDefaultDataset: zod.boolean().optional(),
});

export const TrainModelsResponse = zod.object({
  success: zod.boolean(),
  metrics: zod.array(
    zod.object({
      model: zod.enum(["naive_bayes", "logistic_regression", "svm"]),
      accuracy: zod.number(),
      precision: zod.number(),
      recall: zod.number(),
      f1Score: zod.number(),
      confusionMatrix: zod.object({
        matrix: zod.array(zod.array(zod.number())),
        labels: zod.array(zod.string()),
      }),
      classReport: zod.record(
        zod.string(),
        zod.object({
          precision: zod.number(),
          recall: zod.number(),
          f1Score: zod.number(),
          support: zod.number(),
        }),
      ),
      trainingSamples: zod.number(),
      testSamples: zod.number(),
      trainingTimeMs: zod.number(),
    }),
  ),
  bestModel: zod.enum(["naive_bayes", "logistic_regression", "svm"]),
  totalTimeMs: zod.number(),
});

/**
 * Returns the last-trained model metrics including accuracy, confusion matrix, etc.
 * @summary Get model performance metrics
 */
export const GetModelMetricsResponse = zod.object({
  trained: zod.boolean(),
  metrics: zod.array(
    zod.object({
      model: zod.enum(["naive_bayes", "logistic_regression", "svm"]),
      accuracy: zod.number(),
      precision: zod.number(),
      recall: zod.number(),
      f1Score: zod.number(),
      confusionMatrix: zod.object({
        matrix: zod.array(zod.array(zod.number())),
        labels: zod.array(zod.string()),
      }),
      classReport: zod.record(
        zod.string(),
        zod.object({
          precision: zod.number(),
          recall: zod.number(),
          f1Score: zod.number(),
          support: zod.number(),
        }),
      ),
      trainingSamples: zod.number(),
      testSamples: zod.number(),
      trainingTimeMs: zod.number(),
    }),
  ),
  bestModel: zod.enum(["naive_bayes", "logistic_regression", "svm"]).optional(),
  lastTrainedAt: zod.string().nullish(),
});

/**
 * Run all models on same text and compare results
 * @summary Compare ML models on text
 */
export const CompareModelsBody = zod.object({
  text: zod.string(),
});

export const CompareModelsResponse = zod.object({
  text: zod.string(),
  comparisons: zod.array(
    zod.object({
      label: zod.enum(["positive", "negative", "neutral"]),
      confidence: zod.number(),
      scores: zod.object({
        positive: zod.number(),
        negative: zod.number(),
        neutral: zod.number(),
      }),
      model: zod.enum(["naive_bayes", "logistic_regression", "svm"]),
      processedText: zod.string(),
    }),
  ),
  consensus: zod.enum(["positive", "negative", "neutral"]),
  agreement: zod.number(),
});

/**
 * Generate a downloadable JSON/CSV report of analysis results
 * @summary Generate analysis report
 */
export const GenerateReportBody = zod.object({
  results: zod.array(
    zod.object({
      index: zod.number(),
      text: zod.string(),
      label: zod.enum(["positive", "negative", "neutral"]),
      confidence: zod.number(),
      model: zod.enum(["naive_bayes", "logistic_regression", "svm"]),
    }),
  ),
  format: zod.enum(["json", "csv"]),
  includeMetrics: zod.boolean().optional(),
});

export const GenerateReportResponse = zod.object({
  data: zod.string(),
  filename: zod.string(),
  format: zod.string(),
});

/**
 * Returns word frequencies for a given set of texts and sentiment filter
 * @summary Get word frequency data for word cloud
 */
export const GetWordCloudDataBody = zod.object({
  texts: zod.array(zod.string()),
  sentimentFilter: zod.enum(["positive", "negative", "neutral"]).optional(),
});

export const GetWordCloudDataResponse = zod.object({
  words: zod.array(
    zod.object({
      text: zod.string(),
      value: zod.number(),
    }),
  ),
});