# lib/api-zod/src/generated/types/analyzeBatchResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeBatchResponseSummary } from "./analyzeBatchResponseSummary";
import type { BatchTextResult } from "./batchTextResult";

export interface AnalyzeBatchResponse {
  results: BatchTextResult[];
  summary: AnalyzeBatchResponseSummary;
  processingTimeMs: number;
}