# lib/api-zod/src/generated/types/analyzeTextRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { ModelName } from "./modelName";

export interface AnalyzeTextRequest {
  text: string;
  model?: ModelName;
}