# lib/api-zod/src/generated/types/batchTextResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { ModelName } from "./modelName";
import type { SentimentLabel } from "./sentimentLabel";

export interface BatchTextResult {
  index: number;
  text: string;
  label: SentimentLabel;
  confidence: number;
  model: ModelName;
}