# lib/api-zod/src/generated/types/confusionMatrix.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface ConfusionMatrix {
  matrix: number[][];
  labels: string[];
}