# lib/api-zod/src/generated/types/modelMetrics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { ConfusionMatrix } from "./confusionMatrix";
import type { ModelMetricsClassReport } from "./modelMetricsClassReport";
import type { ModelName } from "./modelName";

export interface ModelMetrics {
  model: ModelName;
  accuracy: number;
  precision: number;
  recall: number;
  f1Score: number;
  confusionMatrix: ConfusionMatrix;
  classReport: ModelMetricsClassReport;
  trainingSamples: number;
  testSamples: number;
  trainingTimeMs: number;
}