# lib/api-zod/src/generated/types/modelMetricsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { ModelMetrics } from "./modelMetrics";
import type { ModelName } from "./modelName";

export interface ModelMetricsResponse {
  trained: boolean;
  metrics: ModelMetrics[];
  bestModel?: ModelName;
  lastTrainedAt?: string | null;
}