# lib/api-zod/src/generated/types/modelName.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */

export type ModelName = (typeof ModelName)[keyof typeof ModelName];

export const ModelName = {
  naive_bayes: "naive_bayes",
  logistic_regression: "logistic_regression",
  svm: "svm",
} as const;