# lib/api-zod/src/generated/types/reportResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface ReportResponse {
  data: string;
  filename: string;
  format: string;
}