# lib/api-zod/src/generated/types/sampleDatasetResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { SampleDatasetResponseSamplesItem } from "./sampleDatasetResponseSamplesItem";

export interface SampleDatasetResponse {
  samples: SampleDatasetResponseSamplesItem[];
  total: number;
  datasetName: string;
}