# lib/api-zod/src/generated/types/sentimentLabel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */

export type SentimentLabel =
  (typeof SentimentLabel)[keyof typeof SentimentLabel];

export const SentimentLabel = {
  positive: "positive",
  negative: "negative",
  neutral: "neutral",
} as const;