# lib/api-zod/src/generated/types/trainResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microtext Sentiment Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { ModelMetrics } from "./modelMetrics";
import type { ModelName } from "./modelName";

export interface TrainResponse {
  success: boolean;
  metrics: ModelMetrics[];
  bestModel: ModelName;
  totalTimeMs: number;
}